Upgraded project file for Xcode 4, fixed warnings.

+++ Classes/Utils.m
@@ -67,7 +67,7 @@ static Utils* sharedInstance = nil;
     return self;
 }
 
-- (unsigned) retainCount {
+- (NSUInteger) retainCount {
     return UINT_MAX;  // denotes an object that cannot be released
 }
